func/requirements.txt
azure-functions>=1.18.0

azure-search-documents==11.7.0b2
azure-core>=1.29.0

python-dotenv>=1.0.0

